backend/image_api/requirements.txt
fastapi
uvicorn
python-dotenv
openai
tika
fastmcp